Create files as local user on linux
Files created by newt were owned by the root user on linux.  This change
runs the container as the caller's user:group.  The docker run command
invokes a wrapper script that creates the user and group, changes the
ownership of the USB devices, and then runs the real newt command.