Update toolchain version
We are not enforcing any semantic versioning with the toolchain, so switched to
a single version number. We didn't want to imply something we weren't doing
with the major.minor.patch version.

Include the toolchain version in the docker image version. The version for the
image is newt_version-toolchain_version.